+++ neo/sys/stub/stub_gl.cpp
@@ -399,6 +399,7 @@ void GLimp_DeactivateContext() {};
 void GLimp_GrabInput(int flags) {};
 bool GLimp_SetSwapInterval( int swapInterval ) { return false; }
 void GLimp_UpdateWindowSize() {}
+bool GLimp_SetWindowResizable( bool enableResizable ) { return false; }
 
 #ifdef _MSC_VER
 #pragma warning(pop)